requirements.txt
agentmail
python-dotenv